requirements.txt
asgiref==3.7.2
Django==4.2.9
djangorestframework==3.14.0
numpy==1.26.3
opencv-python==4.9.0.80
pillow==10.2.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
